Table1 scripts
Create scripts for table 1 and restructure things a bit
